refactor(templates): consolidate create forms into shared TemplateCreateForm (HOL-816)
Extract the three near-duplicate template create forms (org/folder/project)
into a single scope-aware TemplateCreateForm component, mirroring the
BindingForm pattern used by template-policy-bindings. Route files are now
thin wrappers that render breadcrumbs and delegate all form behavior.
Route-level create tests are consolidated into TemplateCreateForm.test.tsx
so the shared behavior is exercised once.
